refactor(ToastNotification): consolidate notifications in different contexts
