+++ packages/components/src/Notification/ToastNotification/ToastNotificationsList/ToastNotificationsList.module.scss
@@ -1,7 +1,7 @@
 @import "~@kaizen/design-tokens/sass/spacing";
 @import "~@kaizen/design-tokens/sass/layout";
 
-.list {
+.toastNotificationsList {
   display: flex;
   flex-direction: column;
   align-items: flex-end;
