--- packages/components/src/Notification/ToastNotification/ToastNotificationsPortal/index.ts
@@ -1,1 +1,0 @@
-export * from "./ToastNotificationsPortal"
